Bug fixes, some major.
Need two different cyclic barriers in pivotpi_servo, one for min
position and one for max.  The flush method in cyclic_barrier was
inadequate and nonstandard, we need to break the cyclic barrier instead
so I added breakAndFlush.  Once cyclic barrier is broken, any thread
that tries to await on it will throw a BrokenBarrierException.  The
flush() method in cyclic barrier will be removed next check-in.